+++ tests/test_json_stream.cpp
@@ -1,0 +1,6 @@
+import std;
+import json;
+
+#include "../test/json_stream_cases.hpp"
+
+int main() { return stream_tests::run(); }